ml_sample/project_template_api/src/pyproject.toml
[tool.poetry]
name = "project_template"
version = "1.0.0"
description = "This is project_template"
authors = ["wararaki"]

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^1.2.1"
pandas = "^1.5.3"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
httpx = "^0.23.3"
pyyaml = "^6.0"
typer = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
